Smoothed the CV input to stop the pitch jumping
The pitch could jump when the ADC had a bit of noise on it. We now only update the pitch if the ADC value changes by more than 8 - which corresponds to 2 Micronotes. There are 4 micronotes per semitone, so basically we need to change by a half a semitone or more. The LFO/vibrato is not effected by this change, and neither is the slide (portamento).