Added hysteresis to the CV input to remove noise.
The ADC could be a bit noisy, so add a half semitone hysteresis to getNote. This stops the note from "warbling" when it is held. The getNote routine uses a lot of RAM outside global variables. When running getNote we consume all 512 bytes of RAM - even a byte more will cause the synth to crash. In the future, we might have to reduce the RAM usage to allow for any new features. Alternatively, decrease the sequencer size.