Fixed a bug in the sequencer, and added a 5th envelope type.
When I cleaned out the sequencer after adding the note hysteresis, I didn't correctly mark the end of the sequence. This crashed the synth as it read memory out of bounds. All fixed now. With the extra space, we now have envelope type 5, which is a symmetric ADSR - which is actually ADSRD so that you can create a resonance envelope. Upon release of the key, we go from the sustain level back to full volume (for that harmonic) at the release speed, then fade away at the decay speed. It makes for quite a nice low pass resonant sweep.