Fixed and modified sequencer
The sequencer had an issue, and in fixing it we have changed how it works. You can now press 6 to start a new sequence. Up to 6 sequences can be defined. Pressing the associated button will jump to that sequence upon finishing the previous one. The is no restriction on the length of sequences other than the notes plus the 5 markers (the first marker is always the first note) is no more than 53 entries (48 notes + 5 markers). This means a single sequence of 53 notes can be defined. This is limited by the available RAM.