Optimized create_wave and added a basic FM engine
By removing 5 multiply operations, we save almost 1/2k of flash. This Enabled me to add a basic FM engine with 2 operators and feedback. It sounds great (in my humble opinion), and allow a larger number of harmonics. 2 operator is quite compatible with the additive synth as the envelopes still control the "pseudo filter" effect. At the end of the day, I just liked the 7:1 chime sound.