Several small bug fixes, improvements and features
I think that about covered it all. A few minor tweaks/bug fixes as well as improving the UI for activating FM. We also now play the timbre while adjusting the envelope "volumes". This makes setting the values much easier. FM feedback now works as expected, so we can get some quite interesting sounds. Back to most of the flash used, so until I find some more space, this is the last improvement for a while.